Rerun configure.
Hacked up a late (post translation to C) tree-shaking pass.  So far it
eliminates only unused structs.  There are a ton of these because:
  * You end up #include'ing a lot transitively
  * Every datatype variant gets translated to a struct
  * Because datatype pointer-types get translated to void*, datatypes
    defined in terms of other datatypes in Cyclone are not defined in
    terms of other datatypes in C.

Anyway, the generated code is over 10% smaller now.

Removing enums is a tad trickier because you have to know the
field->type mapping, so I haven't done that yet.  Also haven't removed
unused externs, which come up with exception tags.

Had to fix bugs in Set, which the new pass uses.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3698 581db07f-6e62-4d40-9733-d432c4eebb07