* Removed alias kinds
  Alias kinds have been replaced completely by aquals(`a). The
following is the mapping from the old to the new

 `a::T(B|A|M) --> `a::(B|A|M) : RESTRICTED <= aquals(`a)

 `r::TR --> `r::R

  T *`U --> T *@aqual(UNIQUE) `H

 To make porting easier the three forms above are translated
automatically to the new representation. So, you will probably still
see uses of `a::TB etc. in several places. This is implemented by the
parser handing off a hint to the tctyp that generates the appropriate
bound. This abuse of the kind notation will be removed in favor of
something that hopefully will be *concise* and make clear that
aliasability is no longer part of the kind.

 The case where the old notation doesn't work at all is in type
instantiation. This change has to be done by hand.

 foo_t<`U> --> foo_t<`H, UNIQUE>

 The runtime system currently supports unique and refcnt'd pointers on
the heap. Plan to change that very soon to get reaps.

 As for testing all these changes ... I've only run it on the stuff in
the tests/ directory -- all of them work except for streambuff, which
worked previously only because of a bug. streambuff uses an
existential type with a refcnt'd pointer inside it ... which is not
allowed. I think it worked earlier because of a bug related to arrays
... but not sure. Need to run this at least on the benchmarks to have
any sort of confidence that I haven't broken something badly.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3875 581db07f-6e62-4d40-9733-d432c4eebb07