Bug fix for setvbuf.  As James Ranson notes, there is a problem with the wrapper for setvbuf.  It should accept a NULL buffer pointer when the mode is _IONFB.
Performance bug fix in assertion widening.  The old code had running
time proportional to the assertion tree, and the new code is proportional
to the dag.  Basically, we associate a set of primitive assertions
with each node corresponding to the weakened version of that node
in a hashtable.  If we encounter that node again, we can immediately
return that set.  

This fixes part of the problem with compiling aprof.cyc, but now it
takes a long time for insert_checks to run because (a) the widening is
a little less precise than the old widening so the simple theorem
prover fails more frequently thus causing us to invoke the Pratt
prover more frequently and (b) the assertions are just so large that
the Pratt prover is choking.  It would be great if we could figure out
some way to avoid blowing up the assertion into a tree again when
doing the Pratt prover.  Otherwise, we're going to have to find some
other way to throttle this bad behavior.  

I'm also still wondering why we're getting such a gigantic assertion.  
Part of the problem is that with a nested "if e1 then s1 else if e2
then s2 else ... else if en then sn else e", by the time we get to
sn, we have an assertion of the form "!e1 && !e1 && ... && !en-1 && en"
but that's a linear thing, not exponential...  On the other hand,
it's DNF could be exponential and on top of that, we definitely have 
a (high-degree) polynomial for the constraint solver so this is probably 
the real problem.  I'm thinking that it might be a good idea to put
an "operation" or "space" throttle on the constraint solver to keep
it from running away like this.  

Added a way to print assertions as dags.  Useful for figuring out
what these giant assertions look like.  

Added support for putting an @assert on a label so that we don't have to
lose all information.  Recall that by default, when we do
vc-generation for a labeled statement "L:S", we reset the
pre-condition to "true".  This means that we don't have to do some
global iteration to handle goto's that are backwards jumps, but means
that you effectively lose all information when you hit a label.  Now,
if the labelled statement S is of the form "@assert(A)" or
"@assert(A);S'", we check that A holds on every path reaching L.

Added support in vcgen to interpret pointer casts so that we better
track when a null or bounds check is needed.  This got rid of a few
more checks in the compiler (but not a lot.)

Added a constraint pass which adds in bounds information for terms
that are thin, possibly-null pointers.  In particular, when t : T*{c},
and we can conclude t != NULL from a graph (i.e., 0 u< t), then we can
add the constraint c <= numelts(t) to the graph.  We delay adding in
this information until other closure passes are run with the hope that
we'll be able to prove t != NULL.  I experimented with adding in the
information early as an explicit assertion (e.g., (t == 0) || 
(c <= numelts(t))) but this results in too many graphs being copied
due to the disjunction.  We need to add a flag to the graphs to
avoid adding this information over and over again.  

Changed constraint solver to iterate the closure passes until there's
no more reductions to perform.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3930 581db07f-6e62-4d40-9733-d432c4eebb07