It would be extremely useful if someone with Makefile wizard status could get "make gprof" and "make PROFILE=X cyclone_src" working.  They almost work, but fail at link time for some strange reason.  I've been able to hack around it manually for now...
Hacking to improve the constraint solver:

* Made graph copying in pratt_prover tail-recursive.

* Modified eq and neq assertion primitives to sort the terms so
that we don't have to worry about commutativity when doing 
simple_prove.  

* Moved widen and simple_prove to assndef and improved the situation
by caching the widened version of an assertion in the hash-cons-table.
That makes it possible to use the widened set when doing
simple_prove(ctxt,a) to avoid re-traversing ctxt.  Rather, we can just
look up the primitive assertions within the weakened version of the
context (which is a set).  This makes simple_prove take time
proportional to the tree representation of a and the dag
representation of ctxt (since computing the set of primitive relations
corresponding to the widen only traverses the dag).  Since ctxt is
usually big, this should help.

* Modified set_dist so that the inner-most loop iterates over the
distance edges of an intermediate node instead of all of the
nodes in the graph.  This seems to cut the running time drastically.
For instance, if I cut out 18 of the "else" clauses in aprof,
then the old compiler (with the changes above) took 45 seconds
whereas the new one takes about 3.  

Alas, this still doesn't make aprof compile in reasonable time.  When
I add in one more else clause, the time goes to 9 sec.  One more and
it goes to 12 sec.  1 more and 14 sec.  1 more and 18 sec.  1 more and
21 sec.  2 more and 33 sec.  2 more and 60 sec (but the conditionals
added are more complicated.)

The problem appears to be that we're adding a zillion constraints.
For instance, one of the calls to the prover generates 4095 distinct
graphs and adds in a total of 577,000 (!) constraints.  The graphs
themselves aren't all that big (roughly 18x18 with I'm speculating
around 50% occupancy).  Anyway, I need to investigate what's causing
the blowup some more.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3937 581db07f-6e62-4d40-9733-d432c4eebb07