* Changed graph construction in pratt_prover.cyc to keep track of assertions that have already been added to the graphs and avoid re-adding them.  This makes graph construction proportional to the dag instead of the tree.
* Changed the way we cut off the search to limit paths.

* Changed the hash-consing for assertions to keep track of the
reduced version of an assertion to avoid recomputing it.  Simplified
all of the hash-consing stuff as well.  

* Changed the way we treat escaping variables and memory so as to
get rid of the ad-hoc Kill(-) constructor for assertions.  Now
we have terms Select(t_mem,t_addr) and Update(t_mem,t_addr,t) as well 
as Addr(x) (where x is a program variable.)  A read of an escaping variable 
x is represented by Select(memory,Addr(x)) where memory is a special 
program variable representing the current memory.  

A read of a non-escaping variable is represented by that variable.  A
write to a non-escaping variable is achieved by substituting a fresh
logic varaible for the program variable.  A write to an escaping
variable (or any location that's not a non-escaping variable) results
in substituting a fresh logic variable for memory.  The constructors
know that Select(Update(m,a,v),a) == a and
Select(Update(m,Addr(x),v),Addr(y)) == Select(m,Addr(y)).  The latter
equation holds since no two distinct program variables are assigned
the same address.  We can do more here (e.g., normalize Update's
by sorting distinct addresses) but it's not clear that it's worth it
for now.  

In the long run, we'll expand the treatment of memory to deal with
more fine-grained interference.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3943 581db07f-6e62-4d40-9733-d432c4eebb07