* I need to fix a bug with tests/zeroterm.cyc...
* Found another bizarre bug in xarray.cyc which I had to work around:
In function Xarray::add, the type-checker seems to assign an array 
type to a pointer.  What's weirder is that the bounds on the array
are "xarr->num_elmts*2" which makes absolutely no sense...  Someone
might want to look into this...

* Added support for projection off of aggregates to vcgen.  This allows 
us to track a lot more values, even through memory and as a result, get 
rid of a few more null and bounds checks.  For instance, doing something 
like: *x->f followed by *x->f, we'll get the NULL-check eliminated as long 
as there's no intervening write to memory.  Here's the effect when
bootstrapping:

  ap:                +2 bounds checks
  graph:             +6 null checks
  printf:            +1 null checks
  set:               +6 null checks
  sexp:              +1 null checks
  string:            +1 null checks
  xarray:            +5 bounds checks
  getopt:           +19 bounds checks, 3 null checks
  queue:             +5 null checks
  splay:             +2 null checks
  absyn:             +1 null checks
  parse:             +3 null checks
  lex:               +1 null checks
  absynpp:          +28 null checks
  absyndump:         +9 null checks
  tcenv:            +11 null checks
  unify:             +2 null checks
  tcutil:           +14 null checks
  tctyp:             +5 null checks
  tcstmt:            +6 null checks
  tcpat:            +21 null checks
  tcexp:            +12 null checks
  evexp:             +4 null checks
  tc:                +2 null checks
  currgn:            +4 null checks
  binding:          +27 null checks
  cf_flowinfo:       +5 null checks
  new_control_flow: +12 null checks
  assndef:           +1 null checks
  vcgen:            +12 null checks
  toc:              +27 null checks
  remove_labels:     +2 null checks
  toseqc:            +2 null checks
  tovc:              +3 null checks
  low_treeshake:     +2 null checks
  interface:         +5 null checks
  cifc:              +1 null checks
  tcdecl:            +2 null checks
  port:              +5 null checks
  cycdoc:            +1 null checks
  buildlib:          +6 null checks

* Track array lengths and not-null when an object is created via a 
comprehension or malloc.  Track not-null on object allocated by new.
These changes tended to slow the compiler down for certain files.
Originally, I was also tracking the results of function calls, but
decided this really slowed things down by too much so took it out for now.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@3953 581db07f-6e62-4d40-9733-d432c4eebb07