Added support for other terms (select, update, proj, etc.) to congruence closure and equal_terms.  Made the loops in congruence closure not quite so bad: First, we specialize the inner loop to each case so that we don't end up pattern matching on the first node each time.  Second, we don't go compare against each node in the inner loop, but rather only start from the node just beyond where the outer loop is.
Added a closure pass to eliminate !Okderef(t) where possible.  In
particular, if we have Okderef(t) <= 0, this encodes that it's not
okay to dereference t.  In turn, this means that for all t1,t2
such that t=t1+t2, it cannot be that t2 U< numelts(t1).  So,
we look for t1 and t2 such that t=t1+t2 and then add the constraint
t2 U>= numelts(t1).  This didn't eliminate any checks.

Added a gaussian equation solver.  The solver finds pairs of terms
that are equal (T1 == T2) and then computes a polynomial representation
of the equation (effectively T1 - T2 == 0, with the left-hand-side
becoming of the form i1*U1 + i2*U2 + ... + in*Un + c.  We then
try to do a form of gaussian elimination to solver for some of
the variables in the set of equations.  We then add in the resulting
equations to the graph.  This allows us to eliminate checks in the
following kinds of code:

void (int @{100} x) {
  int ?p = x;
  p = p + 3;
  @assert(p == x + 3);
  *p;
  p = p - 1;
  @assert(p == x + 2);
  *p;
}

(Unfortunately, the @asserts are still necessary for some reason.)
However, the solver managed to eliminate a few checks in toseqc, 
cf_flowinfo, cycstubs, bitvec, string, sexp, scanf, hashtable, 
streambuff, rc, scheme-2, and scheme.

These changes result in a change of 5.2s for compiling parse_tab.cyc
versus 4.4s for the previous version of the compiler.  

Slight changes to list.cyc and dict.cyc to eliminate a couple of NULL 
and bounds checks.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@4019 581db07f-6e62-4d40-9733-d432c4eebb07