Vcgen now understands patterns and pattern-matching, except that it doesn't yet understand reference patterns into the middle of a struct/ datatype/etc.  This managed to eliminate a few more run-time checks in a number of files (hence the changes.)
By default, we do not track the negation of earlier pattern tests in a
switch, but you can recover the precision by using --allpaths (or the
allpaths pragma.)  This doesn't matter for something like an integer-
switch, except for the default case.  For the richer kinds of patterns,
it loses a fair amount of information.

The reason for not remembering this information by default is that
this slows compile times for the compiler itself (since we use a lot
of pattern matching.)  For instance, the current compile times for
parse_tab and tcutil on my laptop are 5.5s and 2.3s respectively, but
if I turn on --allpaths, parse_tab shoots up to 30s, and tcutil goes
to 11s.

I actually wrote the code twice, once using the decision trees and
once without and found that, somewhat surprisingly, *not* using
the decision trees resulted in faster compile times.  I've left
both hunks of code in vcgen in case we want to re-think things.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@4020 581db07f-6e62-4d40-9733-d432c4eebb07