* Added @ensures clauses to new_string, rnew_string, and rqnew_string so that the caller knows the result is as big as the input integer.
* Fixed one bug that Trevor found -- we weren't constraining existential
variables to rule out Mem kinds in aggregate declarations.  However,
there's still a lurking bug in the code generator which needs to get
tracked down.

* Tried to fix some error message bugs concerning fields in aggregates.
One bug was because we weren't resetting the types that were being
compared for a given failure, which resulted in the strange 
@false != @false sort of error messages.

The other bug had to do with subtyping, and I ended up weakening
subtyping to get this to work out.  When we're doing subtyping, we
pass around a list of subtyping assumptions which get checked first
This makes it possible to do deep, recursive sub-typing.  I'm dropping
the assumptions when dealing with function arguments because (a)
they're contravariant and I'm not sure the assumptions in the context
are valid (they probably are) but more importantly (b) it causes
spurious evars to be constrained to strange types, which makes the
error messages really opaque.  The right fix is for us to have a way
to copy a type, generating fresh constraint variables (evars) so that
we can try something on the copy, and if that succeeds, then do it on
the original.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@4214 581db07f-6e62-4d40-9733-d432c4eebb07