Major bug fix in vcgen.  We were not doing updates to fields of local variables that do not escape correctly.  The updates were done on memory instead of the vardecl map (as if they had escaped.)  This required adding a new term (AggrUpdate(a,i,v)) that represents the update of aggregate term a at index i with value v.  The reduction rules reduce this when possible.
I'm still a little worried about local variables that have a struct
with an array inside them, and it's not clear that we're doing updates
to tagged unions correctly (i.e., updating the tag.)

New flag --unprovable prints out all assertions that we weren't able
to prove and whether or not this was due to limiting the number of
paths explored.  An interesting fact is that for string.cyc, turning
up the number of paths doesn't allow us to prove anything more.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@4293 581db07f-6e62-4d40-9733-d432c4eebb07