Added unify before typecmp in exp_cmp for evexp.cyc. This fixes some of the issues with sizeof_t types.
remove_casts for StmtExp would get stuck in an infinite loop if the
StmtExp didn't end with an expression. Such a StmtExp is illegal in
Cyclone, but is allowed by gcc. So, this turns out to be a problem
when C code is extern C included'd.

ap.c is changed because it looks like one less check is killed by
vcgen. Don't think this is related to the xp.c issue that Trevor just
reported. I tried building this on an x86 and got the same result.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@4520 581db07f-6e62-4d40-9733-d432c4eebb07