missed one line change in last commit.
pratt_prover.cyc is cleaned up a little bit. Now we used
typedef dist_length_t to represent the type of distance in
the grpah. There are a few dependencies on dist_length_t
being signed int in the code:

(1) when we are computing shortest path or when we are looking
for negative cycles in the graph, we use "is_signed_int_overflow"
to detect if the sum of two distances overflowed, and
"is_signed_int_plusminus_overflow" to detect if the computation of
"c1+c2-c3" overflowed.

(2) when we are converting an inequality constraint into a graph
edge, some of the edges might have a weight that is out of the
range of signed int, and we will have to ignore these constraints
for now.


git-svn-id: https://source.seas.harvard.edu/svn/cyclone/trunk/cyclone@4879 581db07f-6e62-4d40-9733-d432c4eebb07